refactor: admin hotel select page 수정
